# ⚠️ This repository is no longer maintained ⚠️
This project has been moved to a new repository. Please visit the new repository for the latest updates, to report issues, or to contribute.

**New Repository:** https://github.com/dooboostore-develop/packages

**New NPM Package:** https://www.npmjs.com/package/@dooboostore/simple-boot-http-server-ssr

While this package may still be available on npm, it will no longer receive updates. Please update your dependencies to point to the new package if available.

---